contrib/QualityInspector/requirements.txt
paddlepaddle-gpu>=2.2.0
paddledet>=2.6
paddleseg>=2.6
visualdl>=2.2.0
opencv-python <= 4.6.0
pycocotools
numpy
PyYAML
tqdm
sklearn
Pillow
matplotlib
scikit-image

# for uad
pandas

